requirements.txt
chromadb
pandas
langchain
tiktoken
numpy
crewai
pypdf
crewai_tools
python-dotenv
google-generativeai
streamlit
matplotlib
fpdf